moving assetfetcher to common
